LayoutTests/platform/ios/imported/w3c/web-platform-tests/screen-orientation/onchange-event-subframe-expected.txt

CONSOLE MESSAGE: Error while parsing the 'sandbox' attribute: 'allow-orientation-lock' is an invalid sandbox flag.

FAIL onchange-event-subframe assert_true: iframe supports 'allow-orientation-lock' expected true got false
PASS Test cross-origin sub-frame receive orientation change events
PASS Check directly that events are fired in right order (from top to bottom)
